% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpdp.R
\name{fastpdp}
\alias{fastpdp}
\alias{fastpdp.default}
\alias{fastpdp.ranger}
\alias{fastpdp.Learner}
\title{Fast PDP}
\usage{
fastpdp(object, ...)

\method{fastpdp}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  w = NULL,
  grid = NULL,
  grid_type = c("fixed", "random"),
  grid_size = 27L,
  trim = c(0.01, 0.99),
  ...
)

\method{fastpdp}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  w = NULL,
  grid = NULL,
  grid_type = c("fixed", "random"),
  grid_size = 27L,
  trim = c(0.01, 0.99),
  ...
)

\method{fastpdp}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  w = NULL,
  grid = NULL,
  grid_type = c("fixed", "random"),
  grid_size = 27L,
  trim = c(0.01, 0.99),
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)}.}

\item{v}{One or more variable names in \code{X} to calculate partial dependence profiles.}

\item{X}{Dataframe or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing \eqn{K \ge 1} numeric predictions per row. Its first argument
represents the model \code{object}, its second argument a data structure like \code{X}.
Additional (named) arguments are passed via \code{...}.
The default, \code{\link[stats:predict]{stats::predict()}}, will work in most cases.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{grid}{Optional evaluation grid. If \code{v} is a single column name, this is a
vector. Otherwise, a matrix or \code{data.frame} with \code{length(v)} columns.
The default (\code{NULL}) will randomly sample rows from \code{X[, v]}.}

\item{grid_type}{Type of grid. If "fixed", will use unique values for discrete
features, and quantile cuts otherwise.}

\item{grid_size}{This is the number of rows sampled randomly from \code{X[, v]}.
The result serves as evaluation \code{grid}. Only used if \code{grid = NULL}.}

\item{trim}{When \code{grid = NULL} and \code{grid_type = "fixed"}, numeric columns are
trimmed first at those quantiles. Set to \code{c(0, 1)} to avoid trimming.}
}
\value{
An object of class "fastpdp" containing
\itemize{
\item \code{pd}: A vector, matrix or \code{data.frame} with PD values. The row order corresponds
to the \code{grid} attached to the output.
\item \code{grid}: A vector, matrix, or \code{data.frame} representing the evaluation grid in
the same row order as \code{pd}. Note that this is a sorted and deduplicated version
of the input \code{grid} (if provided).
\item \code{replications}: How many times each value/row in the output \code{grid} was present
in the input \code{grid}. Like a case weight of the result.
\item \code{v}: The input \code{v}.
}
}
\description{
Fast PDP
}
\section{Methods (by class)}{
\itemize{
\item \code{fastpdp(default)}: Default Kernel SHAP method.

\item \code{fastpdp(ranger)}: PD method for "ranger" models, see Readme for an example.

\item \code{fastpdp(Learner)}: PD method for "mlr3" models, see Readme for an example.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)
fastpdp(fit, v = "Petal.Width", X = iris)
fastpdp(fit, v = "Sepal.Width", X = iris, grid = seq(0, 2.5, by = 0.1))

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
fastpdp(fit, v = "Petal.Width", X = iris)
fastpdp(fit, v = c("Petal.Width", "Species"), X = iris, grid_type = "random")
}
\references{
Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine.
The Annals of Statistics, 29:1189–1232.
}
