% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_importance.R
\name{fx_importance}
\alias{fx_importance}
\alias{fx_importance.default}
\alias{fx_importance.ranger}
\alias{fx_importance.Learner}
\title{PD Importance}
\usage{
fx_importance(object, ...)

\method{fx_importance}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  n_max = 300L,
  out_names = NULL,
  w = NULL,
  sort = TRUE,
  verbose = TRUE,
  ...
)

\method{fx_importance}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  n_max = 300L,
  out_names = NULL,
  w = NULL,
  sort = TRUE,
  verbose = TRUE,
  ...
)

\method{fx_importance}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  n_max = 300L,
  out_names = NULL,
  w = NULL,
  sort = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)}.}

\item{v}{Vector or list of feature names for which PD importance is to be calculated.
If passed as list, \emph{vectors} of feature names are evaluted together.}

\item{X}{Dataframe or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing \eqn{K \ge 1} numeric predictions per row. Its first argument
represents the model \code{object}, its second argument a data structure like \code{X}.
Additional (named) arguments are passed via \code{...}.
The default, \code{\link[stats:predict]{stats::predict()}}, will work in most cases.}

\item{n_max}{If \code{X} has more than \code{n_max} rows, a random sample of \code{n_max} rows is
selected for calculations.}

\item{out_names}{Names of the output columns.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{sort}{Should the result be sortedby importance? Default is \code{TRUE}.}

\item{verbose}{Should a progress bar be shown? Default is \code{TRUE}.}
}
\value{
A \code{data.frame} with one column holding the feature name(s), and K columns with the
PD importance for the K components of the predict function.
}
\description{
The sample variance of the partial dependence function of a feature \eqn{X^{(j)}}
evaluated over a data set \eqn{D} can be used as measure of (main effect) importance,
defined as
\eqn{PDI(j, D) = \textrm{Var}\left(\sum_{i \in D} \textrm{PD}_{j}(x_i^{(j)})\right)}.
Similarly, we can define the PD importance of two or more features together,
measuring their joint effect (main effect plus interaction).
}
\section{Methods (by class)}{
\itemize{
\item \code{fx_importance(default)}: Default method.

\item \code{fx_importance(ranger)}: Method for "ranger" models
.

\item \code{fx_importance(Learner)}: Method for "mlr3" models.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ ., data = iris)
fx_importance(fit, v = names(iris[-1]), X = iris, verbose = FALSE)

v <- list("Sepal.Width", c("Petal.Width", "Petal.Length"), "Species")
fx_importance(fit, v = v, X = iris, verbose = FALSE)

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
fx_importance(fit, v = v, X = iris)

# MODEL THREE: Gamma GLM with log link
fit <- glm(Sepal.Length ~ ., data = iris, family = Gamma(link = log))

fx_importance(fit, v = names(iris[-1]), X = iris, verbose = FALSE)

# On original scale
fx_importance(fit, v = names(iris[-1]), X = iris, verbose = FALSE, type = "response")
}
