% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_pdp.R
\name{fx_pdp}
\alias{fx_pdp}
\alias{fx_pdp.default}
\alias{fx_pdp.ranger}
\alias{fx_pdp.Learner}
\title{Fast PDP}
\usage{
fx_pdp(object, ...)

\method{fx_pdp}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  ...
)

\method{fx_pdp}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  ...
)

\method{fx_pdp}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)}.}

\item{v}{One or more variable names in \code{X} to calculate partial dependence profiles.}

\item{X}{Dataframe or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing \eqn{K \ge 1} numeric predictions per row. Its first argument
represents the model \code{object}, its second argument a data structure like \code{X}.
Additional (named) arguments are passed via \code{...}.
The default, \code{\link[stats:predict]{stats::predict()}}, will work in most cases.}

\item{grid}{Optional evaluation grid. If \code{v} is a single column name, this is a
vector/factor. Otherwise, a matrix or \code{data.frame} with \code{length(v)} columns.}

\item{grid_size}{Determines the grid When \code{grid = NULL}. Character/factor variables
are evaluated at each unique value. A numeric \code{v} with more than \code{grid_size} unique
values is evaluated at \code{grid_size} quantiles. If \code{v} has length \eqn{p > 1},
the \eqn{p}th root of \code{grid_size} is used instead.}

\item{trim}{A vector with two probabilities used to trim non-discrete numeric \code{v}
before applying quantile binning (only if \code{grid = NULL}).
Set to \code{c(0, 1)} to avoid trimming.}

\item{n_max}{If \code{X} has more rows than \code{n_max}, a random sample of \code{n_max} rows is
selected.}

\item{out_names}{Names of the output columns.}

\item{w}{Optional vector of case weights for each row of \code{X}.}
}
\value{
A dataframe with partial dependence per grid value.
}
\description{
Fast PDP
}
\section{Methods (by class)}{
\itemize{
\item \code{fx_pdp(default)}: Default method.

\item \code{fx_pdp(ranger)}: Method for "ranger" models, see Readme for an example.

\item \code{fx_pdp(Learner)}: Method for "mlr3" models, see Readme for an example.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)
pd <- fx_pdp(fit, v = "Petal.Width", X = iris)
pd[1:4, ]

fx_pdp(fit, v = "Petal.Width", X = iris, grid = seq(0, 1, by = 0.5), out_name = "P")
fx_pdp(fit, v = "Petal.Width", X = iris, grid = seq(1, 0, by = -0.5))
fx_pdp(fit, v = "Species", X = iris)

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
fx_pdp(fit, v = "Species", X = iris)
pd <- fx_pdp(fit, v = c("Petal.Width", "Species"), X = iris)
pd[1:4, ]
}
\references{
Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine.
The Annals of Statistics, 29:1189–1232.
}
