% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2_j.R
\name{H2_j}
\alias{H2_j}
\alias{H2_j.default}
\alias{H2_j.interact}
\title{Overall Interaction Strength}
\usage{
H2_j(object, ...)

\method{H2_j}{default}(object, ...)

\method{H2_j}{interact}(
  object,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = Inf,
  eps = 1e-08,
  ...
)
}
\arguments{
\item{object}{Object of class "interact".}

\item{...}{Further parameters passed to predict function (only for default method).}

\item{normalize}{Should statistic be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{sort}{Should results be sorted by the size of the statistic? Default is \code{TRUE}.
Multioutput predictions are sorted by row means.}

\item{top_m}{How many statistics should be shown? By default \code{Inf} (show all).}

\item{eps}{Threshold below which numerator values are set to 0.}
}
\value{
Matrix of interactions statistics (one row per variable, one column per
prediction dimension).
}
\description{
Friedman and Popescu's \eqn{H^2_j} statistics of overall interaction strength per
feature extracted from the result of \code{\link[=interact]{interact()}}, see Details.
}
\details{
The logic of Friedman and Popescu (2008) is as follows:
If there are no interactions involving feature \eqn{x_j}, we can decompose the
(centered) prediction function \eqn{F} into the sum of the (centered) partial
dependence \eqn{F_j} on \eqn{x_j} and the (centered) partial dependence
\eqn{F_{\setminus j}} on all other features \eqn{\mathbf{x}_{\setminus j}}, i.e.,
\deqn{
  F(\mathbf{x}) = F_j(x_j) + F_{\setminus j}(\mathbf{x}_{\setminus j}).
}
Correspondingly, Friedman and Popescu's \eqn{H^2_j} statistic of overall interaction
strength is given by
\deqn{
  H_{j}^2 = \frac{\frac{1}{n} \sum_{i = 1}^n\big[F(\mathbf{x}_i) - 
  \hat F_j(x_{ij}) - \hat F_{\setminus j}(\mathbf{x}_{i\setminus j})
  \big]^2}{\frac{1}{n} \sum_{i = 1}^n\big[F(\mathbf{x}_i)\big]^2}
}
(check \code{\link[=partial_dep]{partial_dep()}} for all definitions).

\strong{Remarks:}
\enumerate{
\item Partial dependence functions (and \eqn{F}) are all centered to
(possibly weighted) mean 0.
\item Partial dependence functions (and \eqn{F}) are evaluated over the data distribution.
This is different to partial dependence plots, where one uses a fixed grid.
\item Weighted versions follow by replacing all arithmetic means by corresponding
weighted means.
\item Multivariate predictions can be treated in a component-wise manner.
\item \eqn{H^2_j = 0} means there are no interactions associated with \eqn{x_j}.
The higher the value, the more prediction variability comes from interactions
with \eqn{x_j}.
\item Since the denominator is the same for all features, the values of the test
statistics can be compared across features.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{H2_j(default)}: Default method of overall interaction strength.

\item \code{H2_j(interact)}: Overall interaction strength from "interact" object.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
H2_j(inter)

# As a barplot
if (requireNamespace("ggplot2", quietly = TRUE)) {
  plot(inter, stat = 1)
}

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
H2_j(inter)

# As a barplot
if (requireNamespace("ggplot2", quietly = TRUE)) {
  plot(inter, stat = 1)
}
}
\references{
Friedman, Jerome H., and Bogdan E. Popescu. \emph{"Predictive Learning via Rule Ensembles."}
The Annals of Applied Statistics 2, no. 3 (2008): 916-54.
}
\seealso{
\code{\link[=interact]{interact()}}, \code{\link[=H2]{H2()}}, \code{\link[=H2_jk]{H2_jk()}}
}
