% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2_jk.R
\name{H2_jk}
\alias{H2_jk}
\alias{H2_jk.default}
\alias{H2_jk.interact}
\title{Pairwise Interaction Strength}
\usage{
H2_jk(object, ...)

\method{H2_jk}{default}(object, ...)

\method{H2_jk}{interact}(
  object,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = 15L,
  eps = 1e-08,
  plot = TRUE,
  fill = "#2b51a1",
  ...
)
}
\arguments{
\item{object}{Object of class "interact".}

\item{...}{Further parameters passed to \code{geom_bar()}.}

\item{normalize}{Should statistic be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{sort}{Should results be sorted by the size of the statistic? Default is \code{TRUE}.
Multioutput predictions are sorted by row means.}

\item{top_m}{How many statistics should be shown? By default \code{15}.
Set to \code{Inf} to show all.}

\item{eps}{Threshold below which numerator values are set to 0.}

\item{plot}{Should results be plotted? Default is \code{TRUE}. Set to \code{FALSE} to get
the results as matrix.}

\item{fill}{Color of bar (only for univariate statistics).}
}
\value{
A "ggplot" object (if \code{plot = TRUE}), or a matrix of statistics
(one row per variable, one column per prediction dimension). If no pairwise
statistics have been calculated, the function returns \code{NULL}.
}
\description{
Friedman and Popescu's statistics of pairwise interaction strength extracted from the
result of \code{\link[=interact]{interact()}}, see Details. By default, the results are plotted as a
barplot. Set \code{plot = FALSE} to get a matrix of values instead.
}
\details{
Following Friedman and Popescu (2008), if there are no interaction effects between
features \eqn{x_j} and \eqn{x_k}, their two-dimensional (centered) partial dependence
function \eqn{F_{jk}} can be written as the sum of the (centered) univariate partial
dependencies \eqn{F_j} and \eqn{F_k}, i.e.,
\deqn{
  F_{jk}(x_j, x_k) = F_j(x_j)+ F_k(x_k).
}
Correspondingly, Friedman and Popescu's \eqn{H_{jk}^2} statistic of pairwise
interaction strength is defined as
\deqn{
  H_{jk}^2 = \frac{A_{jk}}{B_{jk}},
}
where
\deqn{
   A_{jk} = \frac{1}{n} \sum_{i = 1}^n\big[\hat F_{jk}(x_{ij}, x_{ik}) - 
   \hat F_j(x_{ij}) - \hat F_k(x_{ik})\big]^2
}
and
\deqn{
  B_{jk} = \frac{1}{n} \sum_{i = 1}^n\big[\hat F_{jk}(x_{ij}, x_{ik})\big]^2
}
(check \code{\link[=partial_dep]{partial_dep()}} for all definitions).

\strong{Remarks:}
\enumerate{
\item Remarks 1 to 4 of \code{\link[=H2_j]{H2_j()}} also apply here.
\item \eqn{H^2_{jk} = 0} means there are no interaction effects between \eqn{x_j}
and \eqn{x_k}. The larger the value, the more of the joint effect of the two
features comes from the interaction.
\item Since the denominator differs between variable pairs, unlike \eqn{H_j},
this test statistic is difficult to compare between variable pairs.
If both main effects are very weak, a negligible interaction can get a
high \eqn{H^2_{jk}}. Therefore, Friedman and Popescu (2008) suggests to calculate
\eqn{H^2_{jk}} only for \emph{important} variables (see "Modification" below).
}

\strong{Modification}

To be better able to compare pairwise interaction strength across variable pairs,
and to overcome the problem mentioned in the last remark, we suggest as alternative
the unnormalized test statistic on the scale of the predictions,
i.e., \eqn{\sqrt{A_{jk}}}. Set \code{normalize = FALSE} and \code{squared = FALSE} to obtain
this statistic.
Furthermore, we do pairwise calculations not for the most \emph{important} features but
rather for those features with \emph{strongest overall interactions}.
}
\section{Methods (by class)}{
\itemize{
\item \code{H2_jk(default)}: Default pairwise interaction strength.

\item \code{H2_jk(interact)}: Pairwise interaction strength from "interact" object.

}}
\examples{
# MODEL 1: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)

# Proportion of joint effect coming from pairwise interaction
# (for features with strongest overall interactions)
H2_jk(inter)
H2_jk(inter, plot = FALSE)

# Absolute measure as alternative
H2_jk(inter, normalize = FALSE, squared = FALSE)

# MODEL 2: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
H2_jk(inter)
}
\references{
Friedman, Jerome H., and Bogdan E. Popescu. \emph{"Predictive Learning via Rule Ensembles."}
The Annals of Applied Statistics 2, no. 3 (2008): 916-54.
}
\seealso{
\code{\link[=interact]{interact()}}, \code{\link[=H2]{H2()}}, \code{\link[=H2_j]{H2_j()}}
}
