% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{interact}
\alias{interact}
\alias{interact.default}
\alias{interact.ranger}
\alias{interact.Learner}
\title{Calculate Interaction Statistics}
\usage{
interact(object, ...)

\method{interact}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  pairwise_m = 5L,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  ...
)

\method{interact}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  pairwise_m = 5L,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  ...
)

\method{interact}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  pairwise_m = 5L,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)}, for instance
\code{type = "response"} in a \code{\link[=glm]{glm()}} model.}

\item{v}{Vector of feature names.}

\item{X}{A data.frame or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing K >= 1 numeric predictions per row. Its first argument represents the
model \code{object}, its second argument a data structure like \code{X}. Additional arguments
(such as \code{type = "response"} in a GLM) can be passed via \code{...}. The default,
\code{\link[stats:predict]{stats::predict()}}, will work in most cases. Note that column names in a resulting
matrix of predictions will be used as default column names in the results.}

\item{pairwise_m}{Number of features for which pairwise statistics are calculated.
The features are selected based on Friedman and Popescu's overall interaction
strength \eqn{H^2_j} (rowwise maximum in the multivariate case).
Set to \code{length(v)} to not miss any pairwise interaction.
Set to 0 to not calculate any pairwise interaction.}

\item{n_max}{If \code{X} has more than \code{n_max} rows, a random sample of \code{n_max} rows is
selected from \code{X}. In this case, set a random seed for reproducibility.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{verbose}{Should a progress bar be shown? The default is \code{TRUE}.}
}
\value{
An object of class "interact" containing these elements:
\itemize{
\item \code{f}: Matrix with predictions.
\item \code{mean_f2}: (Weighted) mean f^2. Used to normalize most statistics.
\item \code{F_j}: List of matrices, each representing univariable PDs.
\item \code{F_not_j}: List of matrices, each representing the PDs of all variables != j.
\item \code{F_jk}: List of matrices, each representing bivariate PDs.
\item \code{w}: Same as input w.
\item \code{H2_j}: Matrix of Friedman and Popescu's \eqn{H^2_j}.
\item \code{v}: Same as input \code{v}.
\item \code{v_pairwise}: Subset of \code{v} with largest \code{H2_j} used for pairwise calculations.
\item \code{combs}: List of variable pairs for which pairwise PDs are available.
}
}
\description{
Expensive crunching behind interaction statistics such as:
\itemize{
\item Friedman and Popescu's \eqn{H^2_j} of overall interaction strength per feature.
\item Friedman and Popescu's \eqn{H^2_{jk}} of pairwise interaction strengths.
\item Unnormalized pairwise statistic \eqn{H_{jk}}.
\item ...
}
}
\section{Methods (by class)}{
\itemize{
\item \code{interact(default)}: Default interact method.

\item \code{interact(ranger)}: Method for "ranger" models.

\item \code{interact(Learner)}: Method for "mlr3" models.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
inter

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
inter
# MODEL THREE: Gamma GLM with log link
fit <- glm(Sepal.Length ~ ., data = iris, family = Gamma(link = log))

# No interactions for additive features, at least on link scale
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
inter

# On original scale, we have interactions everywhere...
inter <- interact(
  fit, v = names(iris[-1]), X = iris, type = "response", verbose = FALSE
)
inter

}
\references{
Friedman, Jerome H., and Bogdan E. Popescu. \emph{"Predictive Learning via Rule Ensembles."}
The Annals of Applied Statistics 2, no. 3 (2008): 916-54.
}
