% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.fastpdp}
\alias{is.fastpdp}
\title{Check for fastpdp}
\usage{
is.fastpdp(object)
}
\arguments{
\item{object}{An R object.}
}
\value{
\code{TRUE} if \code{object} is of class "fastpdp", and \code{FALSE} otherwise.
}
\description{
Is object of class "fastpdp"?
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
s <- fastpdp(fit, v = "Sepal.Width", X = iris, grid = seq(2, 4.4, by = 0.1))
is.fastpdp(s)
is.fastpdp("a")
}
\seealso{
\code{\link[=fastpdp]{fastpdp()}}
}
