% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_raw.R
\name{make_grid}
\alias{make_grid}
\title{Creates grid of any dimension}
\usage{
make_grid(
  x,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform")
)
}
\arguments{
\item{x}{A vector, matrix, or data.frame to turn into a grid of values.}

\item{grid_size}{Controls the approximate grid size. If \code{X} has p columns, then each
(non-discrete) column will be reduced to about the p-th root of \code{grid_size} values.}

\item{trim}{Non-discrete columns are trimmed at corresponding quantiles.
Set to \code{c(0, 1)} for no trimming.}

\item{strategy}{How should evaluation points of non-discrete columns be found?
Either "quantile" or "uniform".}
}
\value{
A vector, matrix, or data.frame with evaluation points.
}
\description{
Each column of \code{x} is turned independently into a vector of grid values.
Then, all combinations are created with \code{\link[=expand.grid]{expand.grid()}}.
}
\examples{
make_grid(iris$Species)
make_grid(iris[1:2], grid_size = 4)
}
