% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_interaction.R
\name{pd_interaction}
\alias{pd_interaction}
\alias{pd_interaction.default}
\alias{pd_interaction.ranger}
\alias{pd_interaction.Learner}
\title{Fast Friedman's H}
\usage{
pd_interaction(object, ...)

\method{pd_interaction}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  pairwise = FALSE,
  n_max = 200L,
  w = NULL,
  verbose = TRUE,
  ...
)

\method{pd_interaction}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  pairwise = FALSE,
  n_max = 200L,
  w = NULL,
  verbose = TRUE,
  ...
)

\method{pd_interaction}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  pairwise = FALSE,
  n_max = 200L,
  w = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)}.}

\item{v}{Vector of feature names.}

\item{X}{A data.frame or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing K >= 1 numeric predictions per row. Its first argument represents the
model \code{object}, its second argument a data structure like \code{X}. Additional arguments
(such as \code{type = "response"} in a GLM) can be passed via \code{...}. The default,
\code{\link[stats:predict]{stats::predict()}}, will work in most cases. Note that column names in a resulting
matrix of predictions will be used as default column names in the results.}

\item{pairwise}{The default (\code{FALSE}) calculates overall interaction strength per
feature. Set to \code{TRUE} to get \emph{pairwise} statistics (slower).}

\item{n_max}{If \code{X} has more than \code{n_max} rows, a random sample of \code{n_max} rows is
selected for calculations (after determining \code{grid}).}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{verbose}{Should a progress bar be shown? The default is \code{TRUE}.}
}
\value{
An object of class "pd_interaction", containing these elements:
\itemize{
\item \code{num}: Matrix with squared numerator values.
\item \code{denom}: Matrix with squared denominator values.
\item \code{v}: Same as input \code{v}.
\item \code{pairwise}: Same as input \code{pairwise}.
}
}
\description{
Fast Friedman's H
}
\section{Methods (by class)}{
\itemize{
\item \code{pd_interaction(default)}: Default method.

\item \code{pd_interaction(ranger)}: Method for "ranger" models
.

\item \code{pd_interaction(Learner)}: Method for "mlr3" models.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- pd_interaction(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
inter
summary(inter)
summary(inter, normalize = FALSE, squared = TRUE)

inter <- pd_interaction(
  fit, v = names(iris[-1]), X = iris, verbose = FALSE, pairwise = TRUE
)
summary(inter)

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
summary(pd_interaction(fit, v = v, X = iris))
summary(pd_interaction(fit, v = v, X = iris, pairwise = TRUE))

# MODEL THREE: Gamma GLM with log link
fit <- glm(
  Sepal.Length ~ . + Petal.Width:Species, 
  data = iris, 
  family = Gamma(link = log)
)

# No interactions for additive features, at least on link scale
inter <- pd_interaction(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
summary(inter)

# On original scale, we have interactions everywhere...
inter <- pd_interaction(
  fit, v = names(iris[-1]), X = iris, verbose = FALSE, type = "response"
)
summary(inter)

}
\references{
Friedman, J. H. and Popescu, B. E. (2008). "Predictive learning via rule
ensembles." The Annals of Applied Statistics. JSTOR, 916–54.
}
