% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDI2_j.R
\name{PDI2_j}
\alias{PDI2_j}
\alias{PDI2_j.default}
\alias{PDI2_j.interact}
\title{PD Bases Importance (Experimental)}
\usage{
PDI2_j(object, ...)

\method{PDI2_j}{default}(object, ...)

\method{PDI2_j}{interact}(
  object,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = Inf,
  eps = 1e-08,
  plot = TRUE,
  fill = "#2b51a1",
  ...
)
}
\arguments{
\item{object}{Object of class "interact".}

\item{...}{Further parameters passed to \code{geom_bar()}.}

\item{normalize}{Should statistic be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{sort}{Should results be sorted by the size of the statistic? Default is \code{TRUE}.
Multioutput predictions are sorted by row means.}

\item{top_m}{How many statistics should be shown? By default \code{Inf} (show all).}

\item{eps}{Threshold below which numerator values are set to 0.}

\item{plot}{Should results be plotted (default is \code{TRUE}). Set to \code{FALSE} to get
the results as matrix.}

\item{fill}{Color of bar (only for univariate statistics).}
}
\value{
A "ggplot" object (if \code{plot = TRUE}) or a matrix of statistics
(one row per variable, one column per prediction dimension).
}
\description{
Experimental variable importance method based on partial dependence functions.
While related to Greenwell et al., our suggestion measures not only main effect
strength but also interaction effects. It is very closely related to the
\eqn{H^2_j} statistics, see Details.
}
\details{
If \eqn{x_j} has no effects, the (centered) prediction function \eqn{F}
equals the (centered) partial dependence \eqn{F_{\setminus j}} on all other
features \eqn{\mathbf{x}_{\setminus j}}, i.e.,
\deqn{
    F(\mathbf{x}) = F_{\setminus j}(\mathbf{x}_{\setminus j}).
}
Therefore, the following measure of variable importance follows:
\deqn{
  \textrm{PDI}_{j}^2 = \frac{\frac{1}{n} \sum_{i = 1}^n\big[F(\mathbf{x}_i) - 
  \hat F_{\setminus j}(\mathbf{x}_{i\setminus j})\big]^2}{\frac{1}{n} \sum_{i = 1}^n
  \big[F(\mathbf{x}_i)\big]^2}.
}
It differs from \eqn{H^2_j} only by not subtracting the main effect of the \eqn{j}-th
feature in the numerator. It can be read as the proportion of prediction variability
unexplained by all other features. As such, it measures variable importance of
the \eqn{j}-th feature, including its interaction effects (check \code{\link[=partial_dep]{partial_dep()}}
for all definitions).

Remarks 1 to 4 of \code{\link[=H2_j]{H2_j()}} also apply here.
}
\section{Methods (by class)}{
\itemize{
\item \code{PDI2_j(default)}: Default method of PD based feature importance.

\item \code{PDI2_j(interact)}: PD based feature importance from "interact" object.

}}
\examples{
# MODEL 1: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
PDI2_j(inter)
PDI2_j(inter, plot = FALSE)

# MODEL 2: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
PDI2_j(inter)
}
\references{
Greenwell, Brandon M., Bradley C. Boehmke, and Andrew J. McCarthy.
\emph{A Simple and Effective Model-Based Variable Importance Measure.} Arxiv (2018).
}
\seealso{
\code{\link[=interact]{interact()}}, \code{\link[=H2_j]{H2_j()}}
}
