% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_pd.R
\name{pdp_raw}
\alias{pdp_raw}
\title{Barebone Partial Dependence Function}
\usage{
pdp_raw(
  object,
  v,
  X,
  pred_fun,
  grid,
  w = NULL,
  compress_X = TRUE,
  compress_grid = TRUE,
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{v}{Vector of variable name(s) to calculate partial dependence for.}

\item{X}{Dataframe or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing K >= 1 numeric predictions per row. Its first argument represents the
model \code{object}, its second argument a data structure like \code{X}. Additional arguments
(such as \code{type = "response"} in a GLM) can be passed via \code{...}. The default,
\code{\link[stats:predict]{stats::predict()}}, will work in most cases. Note that column names in a resulting
matrix of predictions will be used as default column names in the results.}

\item{grid}{A vector (if \code{length(v) == 1L}), or a matrix/data.frame otherwise.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{compress_X}{If \code{X} has a single non-\code{v} column: should duplicates be removed
and compensated via case weights? (Applied only if >5\% duplicates.)
Default is \code{TRUE}.}

\item{compress_grid}{Should duplicates in \code{grid} be removed, and resulting PDs
mapped back to the original grid index? (Applied only if >5\% duplicates.)
Default is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)}.}
}
\value{
A matrix of partial dependence values (one column per prediction dimension,
one row per grid row).
}
\description{
Creates partial dependence values for a given model, data, and grid.
This is the workhorse function behind \code{\link[=fx_pd]{fx_pd()}}, \code{\link[=fx_interaction]{fx_interaction()}},
and \code{\link[=fx_importance]{fx_importance()}}.

It is fast because:
\enumerate{
\item There is a single call to \code{pred_fun()}.
\item If more than 5\% of grid rows are duplicated, they are dropped from the
calculation. Resulting PD values are mapped back. This is important for
\code{\link[=fx_interaction]{fx_interaction()}} and \code{\link[=fx_importance]{fx_importance()}} that use sampled data as grid.
\item If more than 5\% of non-grid rows are duplicated, these are dropped and
compensated by summing up their case weights.
}
}
\examples{
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
pdp_raw(fit, v = "Petal.Width", X = iris, pred_fun = predict, grid = 1:2)
}
\references{
Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine.
The Annals of Statistics, 29:1189–1232.
}
