% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{plot.hstats_matrix}
\alias{plot.hstats_matrix}
\title{Plots "hstats_matrix" Object}
\usage{
\method{plot}{hstats_matrix}(x, top_m = 15L, fill = "#2b51a1", ...)
}
\arguments{
\item{x}{An object of class "hstats_matrix".}

\item{top_m}{How many rows should be shown? (\code{Inf} to show all.)}

\item{fill}{Color of bar (only for univariate statistics).}

\item{...}{Arguments passed to \link[ggplot2:geom_bar]{ggplot2::geom_bar}.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for objects of class "hstats_matrix".
}
\seealso{
See \code{\link[=perm_importance]{perm_importance()}} for examples.
}
