% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{plot.ice}
\alias{plot.ice}
\title{Plots "ice" Object}
\usage{
\method{plot}{ice}(
  x,
  alpha = 0.2,
  rotate_x = FALSE,
  color = "#2b51a1",
  facet_scales = "free_y",
  ...
)
}
\arguments{
\item{x}{An object of class "ice".}

\item{alpha}{Transparency passed to geometries.}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{color}{Color of lines and points (in case there is no color/fill aesthetic).}

\item{facet_scales}{Value passed to \code{facet_wrap(scales = ...)}.}

\item{...}{Arguments passed to geometries.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for objects of class "ice". Can do (grouped) line plots or
heatmaps.
}
\seealso{
See \code{\link[=ice]{ice()}} for examples.
}
