% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{plot.interact}
\alias{plot.interact}
\title{Plot Interaction Statistics}
\usage{
\method{plot}{interact}(x, stat = 1:2, top_m = 10L, fill = "#2b51a1", ...)
}
\arguments{
\item{x}{An object of class "interact".}

\item{stat}{Which statistic(s) to be shown? Default is \code{1:2}, i.e., show both
\eqn{H^2_j} (1) and \eqn{H^2_{jk}} (2).}

\item{top_m}{Maximum number of rows of results to plot.}

\item{fill}{Color of bars (univariate case).}

\item{...}{Further arguments passed to statistics \code{\link[=H2_j]{H2_j()}} and \code{\link[=H2_jk]{H2_jk()}}.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for object of class "interact".
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  # MODEL ONE: Linear regression
  fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
  inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
  plot(inter)

  # MODEL TWO: Multi-response linear regression
  fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
  v <- c("Petal.Length", "Petal.Width", "Species")
  inter <- interact(fit, v = v, X = iris, verbose = FALSE)
  plot(inter, stat = 1)
}

}
\seealso{
\code{\link[=interact]{interact()}}
}
