% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.fastpdp}
\alias{print.fastpdp}
\title{Print Method}
\usage{
\method{print}{fastpdp}(x, ...)
}
\arguments{
\item{x}{An object of class "fastpdp".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Prints the first two rows of the matrix (or matrices) of SHAP values.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
s <- fastpdp(fit, v = "Species", X = iris)
s
}
\seealso{
\code{\link[=fastpdp]{fastpdp()}}
}
