% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_importance.R
\name{print.fx_importance}
\alias{print.fx_importance}
\title{Importance Print}
\usage{
\method{print}{fx_importance}(x, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Currently unused.}
}
\value{
Invisibly, \code{x} is returned.
}
\description{
Print function for result of \code{\link[=fx_importance]{fx_importance()}}.
}
\seealso{
\code{\link[=fx_importance]{fx_importance()}}
}
