% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_interaction.R
\name{print.fx_interaction}
\alias{print.fx_interaction}
\title{Interaction Print}
\usage{
\method{print}{fx_interaction}(x, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Currently unused.}
}
\value{
Invisibly, \code{x} is returned.
}
\description{
Print function for result of \code{\link[=fx_interaction]{fx_interaction()}}.
}
\seealso{
\code{\link[=fx_interaction]{fx_interaction()}}
}
