% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_pd.R
\name{print.fx_pd}
\alias{print.fx_pd}
\title{PD Print}
\usage{
\method{print}{fx_pd}(x, which = 1L, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Currently unused.}
}
\value{
Invisibly, \code{x} is returned.
}
\description{
Print function for result of \code{\link[=fx_pd]{fx_pd()}}.
}
\seealso{
\code{\link[=fx_pd]{fx_pd()}}
}
