% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstats.R
\name{print.hstats}
\alias{print.hstats}
\title{Print Method}
\usage{
\method{print}{hstats}(x, ...)
}
\arguments{
\item{x}{An object of class "hstats".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for object of class "hstats". Shows \eqn{H^2} statistic.
}
\seealso{
See \code{\link[=hstats]{hstats()}} for examples.
}
