% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{print.hstats_matrix}
\alias{print.hstats_matrix}
\title{Prints "hstats_matrix" Object}
\usage{
\method{print}{hstats_matrix}(x, top_m = Inf, ...)
}
\arguments{
\item{x}{An object of class "hstats_matrix".}

\item{top_m}{Number of rows to print.}

\item{...}{Currently not used.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for object of class "hstats_matrix".
}
