% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstats.R
\name{print.hstats_summary}
\alias{print.hstats_summary}
\title{Print Method}
\usage{
\method{print}{hstats_summary}(x, ...)
}
\arguments{
\item{x}{An object of class "hstats_summary".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for object of class "hstats_summary".
}
\seealso{
See \code{\link[=hstats]{hstats()}} for examples.
}
