% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{print.ice}
\alias{print.ice}
\title{Prints "ice" Object}
\usage{
\method{print}{ice}(x, n = 3L, ...)
}
\arguments{
\item{x}{An object of class "ice".}

\item{n}{Number of rows of partial dependencies to show.}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for object of class "ice".
}
\seealso{
See \code{\link[=ice]{ice()}} for examples.
}
