% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{print.interact}
\alias{print.interact}
\title{Print Method}
\usage{
\method{print}{interact}(x, ...)
}
\arguments{
\item{x}{An object of class "interact".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for object of class "interact".
}
\examples{
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
inter 
}
\seealso{
\code{\link[=interact]{interact()}}
}
