% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dep.R
\name{print.partial_dep}
\alias{print.partial_dep}
\title{Prints "partial_dep" Object}
\usage{
\method{print}{partial_dep}(x, n = 3L, ...)
}
\arguments{
\item{x}{An object of class "partial_dep".}

\item{n}{Number of rows of partial dependencies to show.}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for object of class "partial_dep".
}
\seealso{
See \code{\link[=partial_dep]{partial_dep()}} for examples.
}
