% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{print.pd}
\alias{print.pd}
\title{Print Method}
\usage{
\method{print}{pd}(x, ...)
}
\arguments{
\item{x}{An object of class "pd".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for object of class "pd".
}
\seealso{
See \code{\link[=partial_dep]{partial_dep()}} for examples.
}
