% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_importance.R
\name{print.pd_importance}
\alias{print.pd_importance}
\title{Importance Print}
\usage{
\method{print}{pd_importance}(x, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Currently unused.}
}
\value{
Invisibly, \code{x} is returned.
}
\description{
Print function for result of \code{\link[=pd_importance]{pd_importance()}}.
}
\seealso{
\code{\link[=pd_importance]{pd_importance()}}
}
