% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_interaction.R
\name{print.pd_interaction}
\alias{print.pd_interaction}
\title{Interaction Print}
\usage{
\method{print}{pd_interaction}(x, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Currently unused.}
}
\value{
Invisibly, \code{x} is returned.
}
\description{
Print function for result of \code{\link[=pd_interaction]{pd_interaction()}}.
}
\seealso{
\code{\link[=pd_interaction]{pd_interaction()}}
}
