% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_profiles.R
\name{print.pd_profiles}
\alias{print.pd_profiles}
\title{pd_profiles Print}
\usage{
\method{print}{pd_profiles}(x, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Currently unused.}
}
\value{
Invisibly, \code{x} is returned.
}
\description{
Print function for result of \code{\link[=pd_profiles]{pd_profiles()}}.
}
\seealso{
\code{\link[=pd_profiles]{pd_profiles()}}
}
