% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_importance.R
\name{print.perm_importance}
\alias{print.perm_importance}
\title{Print Method}
\usage{
\method{print}{perm_importance}(x, ...)
}
\arguments{
\item{x}{An object of class "perm_importance".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for object of class "perm_importance".
}
\seealso{
See \code{\link[=perm_importance]{perm_importance()}} for examples.
}
