% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstats.R
\name{print.summary_hstats}
\alias{print.summary_hstats}
\title{Print Method}
\usage{
\method{print}{summary_hstats}(x, ...)
}
\arguments{
\item{x}{An object of class "summary_hstats".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for object of class "summary_hstats".
}
\seealso{
See \code{\link[=hstats]{hstats()}} for examples.
}
