% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_interaction.R
\name{summary.fx_interaction}
\alias{summary.fx_interaction}
\title{Interaction Summary}
\usage{
\method{summary}{fx_interaction}(
  object,
  normalize = TRUE,
  squared = FALSE,
  sort = TRUE,
  top_m = Inf,
  out_names = NULL,
  eps = 1e-08,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Object to summarize.}

\item{normalize}{Should explained variances be normalized? Default is \code{TRUE}.}

\item{squared}{Should squared statistics be returned? Default is \code{FALSE}.}

\item{sort}{Should result be sorted? Default is \code{TRUE}.}

\item{top_m}{How many rows should be shown? By default \code{Inf} (show all).}

\item{out_names}{Optional names of the output columns corresponding to the
K-dimensional predictions.}

\item{eps}{Threshold below which num values are set to 0.}

\item{verbose}{Should description be printed or not? Default is \code{TRUE}.}

\item{...}{Currently unused.}
}
\value{
Matrix with statistics (one column per prediction dimension).
}
\description{
Uses the results of \code{\link[=fx_interaction]{fx_interaction()}} to calculate different versions of
Friedman's H, see \code{\link[=fx_interaction]{fx_interaction()}}.
}
\seealso{
\code{\link[=fx_interaction]{fx_interaction()}}
}
