% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstats.R
\name{summary.hstats}
\alias{summary.hstats}
\title{Summary Method}
\usage{
\method{summary}{hstats}(
  object,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = Inf,
  zero = TRUE,
  eps = 1e-08,
  ...
)
}
\arguments{
\item{object}{Object of class "hstats".}

\item{normalize}{Should statistics be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{sort}{Should results be sorted? Default is \code{TRUE}.
(Multioutput is sorted by row means.)}

\item{top_m}{How many rows should be shown? (\code{Inf} to show all.)}

\item{zero}{Should rows with all 0 be shown? Default is \code{TRUE}.}

\item{eps}{Threshold below which numerator values are set to 0.}

\item{...}{Currently not used.}
}
\value{
An object of class "summary_hstats" representing a named list with statistics
"h2", "h2_overall", "h2_pairwise", "h2_threeway", and the input flag "normalize".
Statistics that equal \code{NULL} are omitted from the list.
}
\description{
Summary method for "hstats" object. Note that \eqn{H^2} is not affected by
the arguments \code{normalize} and \code{squared}.
}
\seealso{
See \code{\link[=hstats]{hstats()}} for examples.
}
