% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{summary.interact}
\alias{summary.interact}
\title{Summary Method}
\usage{
\method{summary}{interact}(object, n = 5L, ...)
}
\arguments{
\item{object}{An object of class "interact".}

\item{n}{Maximum number of rows of results to print.s}

\item{...}{Further arguments passed from other methods.}
}
\value{
A list of resulting matrices.
}
\description{
Summary method for "interact" object.
}
\examples{
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
summary(inter)
}
\seealso{
\code{\link[=interact]{interact()}}
}
