% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_importance.R
\name{summary.pd_importance}
\alias{summary.pd_importance}
\title{Importance Summary}
\usage{
\method{summary}{pd_importance}(
  object,
  sort = TRUE,
  top_m = Inf,
  out_names = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Object to summarize.}

\item{sort}{Should result be sorted? Default is \code{TRUE}.}

\item{top_m}{How many rows should be shown? By default \code{Inf} (show all).}

\item{out_names}{Optional names of the output columns corresponding to the
K-dimensional predictions.}

\item{verbose}{Should description be printed or not? Default is \code{TRUE}.}

\item{...}{Currently unused.}
}
\value{
Matrix with statistics (one column per prediction dimension).
}
\description{
Extracts the results of \code{\link[=pd_importance]{pd_importance()}}.
}
\seealso{
\code{\link[=pd_importance]{pd_importance()}}
}
