% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_profiles.R
\name{summary.pd_profiles}
\alias{summary.pd_profiles}
\title{pd_profiles Summary}
\usage{
\method{summary}{pd_profiles}(object, which = 1L, out_names = NULL, ...)
}
\arguments{
\item{object}{Object to summarize.}

\item{which}{Name or position (within \code{v}) of feature to show partial dependence.}

\item{out_names}{Optional names of the output columns corresponding to the
K-dimensional predictions.}

\item{...}{Currently unused.}
}
\value{
data.frame with evaluation grid and PD values (one column per prediction).
}
\description{
Uses the results of \code{\link[=pd_profiles]{pd_profiles()}} to show PD values for a given variable.
}
\seealso{
\code{\link[=pd_profiles]{pd_profiles()}}
}
