% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_importance.R
\name{summary.perm_importance}
\alias{summary.perm_importance}
\title{Summary Method}
\usage{
\method{summary}{perm_importance}(
  object,
  normalize = FALSE,
  sort = TRUE,
  top_m = 15L,
  agg_cols = c("sum", "mean", "max", "no"),
  err_type = c("se", "sd", "no"),
  ...
)
}
\arguments{
\item{object}{An object of class "perm_importance".}

\item{normalize}{Should importance statistics be divided by performance?
Default is \code{FALSE}. Normalization is applied after collapsing multivariate losses.}

\item{sort}{Should results be sorted by importance? Default is \code{TRUE}.
The multivariate case is sorted by row sums.}

\item{top_m}{Of how many features should importance values be shown?
Default is \code{Inf}, i.e., show all.}

\item{agg_cols}{Name of function used to collapse multivariate losses.
Default is \code{"sum"}. Set to \code{"no"} for no aggregation. Other options are "mean" and
"max". This step is applied before normalization.}

\item{err_type}{Should standard errors ("se", default) or standard deviations ("sd")
be calculated? Only relevant if \code{perms > 1}. Set to "no" to hide errors.}

\item{...}{Currently not used.}
}
\value{
A list with two matrices: \code{imp} contains the importance values per feature,
while \code{err} represents the uncertainty of these values.
}
\description{
Summary method for "perm_importance" object. By default shows the top 15 feature
importances collapsed over repetitions (and loss dimensions).
}
\seealso{
See \code{\link[=perm_importance]{perm_importance()}} for examples.
}
