% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_interaction_strength.R
\name{total_interaction_strength}
\alias{total_interaction_strength}
\alias{total_interaction_strength.default}
\alias{total_interaction_strength.ranger}
\alias{total_interaction_strength.Learner}
\alias{total_interaction_strength.interact}
\title{Total Interaction Strength}
\usage{
total_interaction_strength(object, ...)

\method{total_interaction_strength}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  eps = 1e-08,
  ...
)

\method{total_interaction_strength}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  eps = 1e-08,
  ...
)

\method{total_interaction_strength}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  eps = 1e-08,
  ...
)

\method{total_interaction_strength}{interact}(
  object,
  normalize = TRUE,
  squared = TRUE,
  eps = 1e-08,
  ...
)
}
\arguments{
\item{object}{Object of class "interact", or a model object.}

\item{...}{Further parameters passed to predict function (only for default method).}

\item{v}{Vector of feature names.}

\item{X}{A data.frame or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing K >= 1 numeric predictions per row. Its first argument represents the
model \code{object}, its second argument a data structure like \code{X}. Additional arguments
(such as \code{type = "response"} in a GLM) can be passed via \code{...}. The default,
\code{\link[stats:predict]{stats::predict()}}, will work in most cases. Note that column names in a resulting
matrix of predictions will be used as default column names in the results.}

\item{n_max}{If \code{X} has more than \code{n_max} rows, a random sample of \code{n_max} rows is
selected from \code{X}. In this case, set a random seed for reproducibility.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{verbose}{Should a progress bar be shown? The default is \code{TRUE}.}

\item{normalize}{Should statistic be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{eps}{Threshold below which numerator values are set to 0.}
}
\value{
Vector of total interaction strength (one value per prediction dimension).
}
\description{
Total Interaction Strength
}
\section{Methods (by class)}{
\itemize{
\item \code{total_interaction_strength(default)}: Default method of total interaction strength.

\item \code{total_interaction_strength(ranger)}: Total interaction strength from "ranger" models.

\item \code{total_interaction_strength(Learner)}: Total interaction strength from "mlr3" models.

\item \code{total_interaction_strength(interact)}: Total interaction strength from
"interact".

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
total_interaction_strength(inter)

\dontrun{
total_interaction_strength(fit, v = names(iris[-1]), X = iris, verbose = FALSE)

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
total_interaction_strength(inter)
total_interaction_strength(fit, v = v, X = iris, verbose = FALSE)

# MODEL THREE: Linear regression
fit <- lm(Sepal.Length ~ ., data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
total_interaction_strength(inter)
total_interaction_strength(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
}
}
