% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_friedmans_h.R
\name{fx_friedmans_h}
\alias{fx_friedmans_h}
\alias{fx_friedmans_h.default}
\alias{fx_friedmans_h.ranger}
\alias{fx_friedmans_h.Learner}
\title{Friedman's H}
\usage{
fx_friedmans_h(object, ...)

\method{fx_friedmans_h}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  normalize = TRUE,
  take_sqrt = TRUE,
  grid_size = 50L,
  n_max = 200L,
  out_names = NULL,
  w = NULL,
  verbose = TRUE,
  ...
)

\method{fx_friedmans_h}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  ...
)

\method{fx_friedmans_h}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)}.}

\item{v}{Vector of variables names to calculate Friedman's H for.}

\item{X}{Dataframe or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing \eqn{K \ge 1} numeric predictions per row. Its first argument
represents the model \code{object}, its second argument a data structure like \code{X}.
Additional (named) arguments are passed via \code{...}.
The default, \code{\link[stats:predict]{stats::predict()}}, will work in most cases.}

\item{grid_size}{Determines the grid When \code{grid = NULL}. Character/factor variables
are evaluated at each unique value. A numeric \code{v} with more than \code{grid_size} unique
values is evaluated at \code{grid_size} quantiles. If \code{v} has length \eqn{p > 1},
the \eqn{p}th root of \code{grid_size} is used instead.}

\item{n_max}{If \code{X} has more rows than \code{n_max}, a random sample of \code{n_max} rows is
selected.}

\item{out_names}{Names of the output columns.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{verbose}{Should a progress bar be shown? Default is \code{TRUE}.}

\item{grid}{Optional evaluation grid. If \code{v} is a single column name, this is a
vector/factor. Otherwise, a matrix or \code{data.frame} with \code{length(v)} columns.}

\item{trim}{A vector with two probabilities used to trim non-discrete numeric \code{v}
before applying quantile binning (only if \code{grid = NULL}).
Set to \code{c(0, 1)} to avoid trimming.}
}
\value{
A data.frame
}
\description{
Friedman's H
}
\section{Methods (by class)}{
\itemize{
\item \code{fx_friedmans_h(default)}: Default method.

\item \code{fx_friedmans_h(ranger)}: Method for "ranger" models, see Readme for an example.

\item \code{fx_friedmans_h(Learner)}: Method for "mlr3" models, see Readme for an example.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
fx_friedmans_h(fit, v = c("Petal.Width", "Species"), X = iris)

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
h <- fx_friedmans_h(fit, v = colnames(iris[3:5]), X = iris)
h
fx_friedmans_h(fit, v = c("Petal.Width", "Species"), X = iris)
}
\references{
Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine.
The Annals of Statistics, 29:1189–1232.
}
