% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2.R
\name{H2}
\alias{H2}
\alias{H2.default}
\alias{H2.ranger}
\alias{H2.Learner}
\alias{H2.interact}
\title{Total Interaction Strength}
\usage{
H2(object, ...)

\method{H2}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  eps = 1e-08,
  ...
)

\method{H2}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  eps = 1e-08,
  ...
)

\method{H2}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  eps = 1e-08,
  ...
)

\method{H2}{interact}(object, normalize = TRUE, squared = TRUE, eps = 1e-08, ...)
}
\arguments{
\item{object}{Object of class "interact", or a model object.}

\item{...}{Further parameters passed to predict function (only for default method).}

\item{v}{Vector of feature names.}

\item{X}{A data.frame or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing K >= 1 numeric predictions per row. Its first argument represents the
model \code{object}, its second argument a data structure like \code{X}. Additional arguments
(such as \code{type = "response"} in a GLM) can be passed via \code{...}. The default,
\code{\link[stats:predict]{stats::predict()}}, will work in most cases. Note that column names in a resulting
matrix of predictions will be used as default column names in the results.}

\item{n_max}{If \code{X} has more than \code{n_max} rows, a random sample of \code{n_max} rows is
selected from \code{X}. In this case, set a random seed for reproducibility.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{verbose}{Should a progress bar be shown? The default is \code{TRUE}.}

\item{normalize}{Should statistic be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{eps}{Threshold below which numerator values are set to 0.}
}
\value{
Vector of total interaction strength (one value per prediction dimension).
}
\description{
Proportion of prediction variability coming from interactions,
ideally calculated from the result of \code{\link[=interact]{interact()}}.
}
\details{
The idea is as follows: if the model is additive in all features,
then the (centered) prediction function \eqn{F} equals the sum of the (centered)
partial dependence functions \eqn{F_j(x_j)}, i.e.,
\deqn{
  F(\mathbf{x}) = \sum_{j}^{p} F_j(x_j)
}
(check \code{\link[=pd]{pd()}} for all definitions).
To measure the relative amount of variability explained by all interactions,
we can therefore study the test statistic of total interaction strength
\deqn{
  H^2 = \frac{\frac{1}{n} \sum_{i = 1}^n \big[F(\mathbf{x}_i) - 
  \sum_{j = 1}^p\hat F_j(x_{ij})\big]^2}{\frac{1}{n} 
  \sum_{i = 1}^n\big[F(\mathbf{x}_i)\big]^2}.
}
It equals the variability of the predictions unexplained by the main effects.
A value of 0 would mean there are no interaction effects at all.
}
\section{Methods (by class)}{
\itemize{
\item \code{H2(default)}: Default method of total interaction strength.

\item \code{H2(ranger)}: Total interaction strength from "ranger" models.

\item \code{H2(Learner)}: Total interaction strength from "mlr3" models.

\item \code{H2(interact)}: Total interaction strength from "interact" object.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
H2(inter)

\dontrun{
H2(fit, v = names(iris[-1]), X = iris, verbose = FALSE)

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
H2(inter)
H2(fit, v = v, X = iris, verbose = FALSE)

# MODEL THREE: No interactions
fit <- lm(Sepal.Length ~ ., data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
H2(inter)
}
}
