% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2.R
\name{H2}
\alias{H2}
\alias{H2.default}
\alias{H2.interact}
\title{Total Interaction Strength}
\usage{
H2(object, ...)

\method{H2}{default}(object, ...)

\method{H2}{interact}(object, normalize = TRUE, squared = TRUE, eps = 1e-08, ...)
}
\arguments{
\item{object}{Object of class "interact".}

\item{...}{Currently unused.}

\item{normalize}{Should statistic be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{eps}{Threshold below which numerator values are set to 0.}
}
\value{
Vector of total interaction strength (one value per prediction dimension).
}
\description{
Proportion of prediction variability unexplained by main effects of \code{v}
(extracted from the result of \code{\link[=interact]{interact()}}), see Details.
}
\details{
If the model is additive in all features, then the (centered) prediction
function \eqn{F} equals the sum of the (centered) partial dependence
functions \eqn{F_j(x_j)}, i.e.,
\deqn{
  F(\mathbf{x}) = \sum_{j}^{p} F_j(x_j)
}
(check \code{\link[=partial_dep]{partial_dep()}} for all definitions).
To measure the relative amount of variability unexplained by all main effects,
we can therefore study the test statistic of total interaction strength
\deqn{
  H^2 = \frac{\frac{1}{n} \sum_{i = 1}^n \big[F(\mathbf{x}_i) - 
  \sum_{j = 1}^p\hat F_j(x_{ij})\big]^2}{\frac{1}{n} 
  \sum_{i = 1}^n\big[F(\mathbf{x}_i)\big]^2}.
}
A value of 0 means there are no interaction effects at all.

In Żółkowski et al. (2023), \eqn{1 - H^2} is called \emph{additivity index}.
A similar measure using accumulated local effects is discussed in Molnar (2020).
}
\section{Methods (by class)}{
\itemize{
\item \code{H2(default)}: Default method of total interaction strength.

\item \code{H2(interact)}: Total interaction strength from "interact" object.

}}
\examples{
# MODEL 1: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
H2(inter)

# MODEL 2: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
H2(inter)

# MODEL 3: No interactions
fit <- lm(Sepal.Length ~ ., data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
H2(inter)
}
\references{
\enumerate{
\item Żółkowski, Artur, Mateusz Krzyziński, and Paweł Fijałkowski.
\emph{Methods for extraction of interactions from predictive models.}
Undergraduate thesis. Faculty of Mathematics and Information Science,
Warsaw University of Technology (2023).
\item Molnar, Christoph, Giuseppe Casalicchio, and Bernd Bischl".
\emph{Quantifying Model Complexity via Functional Decomposition for Better Post-hoc Interpretability},
in Machine Learning and Knowledge Discovery in Databases,
Springer International Publishing (2020): 193-204.
}
}
\seealso{
\code{\link[=interact]{interact()}}, \code{\link[=H2_overall]{H2_overall()}}, \code{\link[=H2_pairwise]{H2_pairwise()}}, \code{\link[=H2_threeway]{H2_threeway()}}
}
