% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2.R
\name{H2}
\alias{H2}
\alias{H2.default}
\alias{H2.interact}
\title{Total Interaction Strength}
\usage{
H2(object, ...)

\method{H2}{default}(object, ...)

\method{H2}{interact}(object, normalize = TRUE, squared = TRUE, eps = 1e-08, ...)
}
\arguments{
\item{object}{Object of class "interact".}

\item{...}{Further parameters passed to predict function (only for default method).}

\item{normalize}{Should statistic be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{eps}{Threshold below which numerator values are set to 0.}
}
\value{
Vector of total interaction strength (one value per prediction dimension).
}
\description{
Proportion of prediction variability coming from interactions extracted from the
result of \code{\link[=interact]{interact()}}, see Details.
}
\details{
If the model is additive in all features, then the (centered) prediction
function \eqn{F} equals the sum of the (centered) partial dependence
functions \eqn{F_j(x_j)}, i.e.,
\deqn{
  F(\mathbf{x}) = \sum_{j}^{p} F_j(x_j)
}
(check \code{\link[=partial_dep]{partial_dep()}} for all definitions).
To measure the relative amount of variability explained by all interactions,
we can therefore study the test statistic of total interaction strength
\deqn{
  H^2 = \frac{\frac{1}{n} \sum_{i = 1}^n \big[F(\mathbf{x}_i) - 
  \sum_{j = 1}^p\hat F_j(x_{ij})\big]^2}{\frac{1}{n} 
  \sum_{i = 1}^n\big[F(\mathbf{x}_i)\big]^2}.
}
It equals the variability of the predictions unexplained by the main effects.
A value of 0 would mean there are no interaction effects at all.
}
\section{Methods (by class)}{
\itemize{
\item \code{H2(default)}: Default method of total interaction strength.

\item \code{H2(interact)}: Total interaction strength from "interact" object.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
H2(inter)

\dontrun{
H2(fit, v = names(iris[-1]), X = iris, verbose = FALSE)

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
H2(inter)

# MODEL THREE: No interactions
fit <- lm(Sepal.Length ~ ., data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
H2(inter)
}
}
\seealso{
\code{\link[=interact]{interact()}}, \code{\link[=H2_j]{H2_j()}}, \code{\link[=H2_jk]{H2_jk()}}
}
