% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2_overall.R
\name{H2_overall}
\alias{H2_overall}
\alias{H2_overall.default}
\alias{H2_overall.ranger}
\alias{H2_overall.Learner}
\alias{H2_overall.interact}
\title{Overall Interaction Strength}
\usage{
H2_overall(object, ...)

\method{H2_overall}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = Inf,
  eps = 1e-08,
  ...
)

\method{H2_overall}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = Inf,
  eps = 1e-08,
  ...
)

\method{H2_overall}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = Inf,
  eps = 1e-08,
  ...
)

\method{H2_overall}{interact}(
  object,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = Inf,
  eps = 1e-08,
  ...
)
}
\arguments{
\item{object}{Object of class "interact", or a model object.}

\item{...}{Further parameters passed to predict function (only for default method).}

\item{v}{Vector of feature names.}

\item{X}{A data.frame or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing K >= 1 numeric predictions per row. Its first argument represents the
model \code{object}, its second argument a data structure like \code{X}. Additional arguments
(such as \code{type = "response"} in a GLM) can be passed via \code{...}. The default,
\code{\link[stats:predict]{stats::predict()}}, will work in most cases. Note that column names in a resulting
matrix of predictions will be used as default column names in the results.}

\item{n_max}{If \code{X} has more than \code{n_max} rows, a random sample of \code{n_max} rows is
selected from \code{X}. In this case, set a random seed for reproducibility.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{verbose}{Should a progress bar be shown? The default is \code{TRUE}.}

\item{normalize}{Should statistic be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{sort}{Should results be sorted by the size of the statistic? Default is \code{TRUE}.
Multioutput predictions are sorted by row means.}

\item{top_m}{How many statistics should be shown? By default \code{Inf} (show all).}

\item{eps}{Threshold below which numerator values are set to 0.}
}
\value{
Matrix of interactions statistics (one row per variable, one column per
prediction dimension).
}
\description{
Friedman and Popescu's H^2_j of overall interaction strength.
}
\section{Methods (by class)}{
\itemize{
\item \code{H2_overall(default)}: Default method of overall interaction strength.

\item \code{H2_overall(ranger)}: Overall interaction strength from "ranger" models.

\item \code{H2_overall(Learner)}: Overall interaction strength from "mlr3" models.

\item \code{H2_overall(interact)}: Overall interaction strength from "interact" object.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
H2_overall(inter)

\dontrun{
H2_overall(fit, v = names(iris[-1]), X = iris, verbose = FALSE)

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
H2_overall(inter)
H2_overall(fit, v = v, X = iris, verbose = FALSE)
}
}
\references{
Friedman, Jerome H., and Bogdan E. Popescu. "Predictive Learning via Rule Ensembles."
The Annals of Applied Statistics 2, no. 3 (2008): 916-54.
}
