% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Friedmans_H.R
\name{H2_pairwise}
\alias{H2_pairwise}
\alias{H2_pairwise.default}
\alias{H2_pairwise.interaction_statistics}
\title{Pairwise Interaction Strength}
\usage{
H2_pairwise(object, ...)

\method{H2_pairwise}{default}(object, ...)

\method{H2_pairwise}{interaction_statistics}(
  object,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = Inf,
  eps = 1e-08,
  ...
)
}
\arguments{
\item{object}{Object of class "interaction_statistics".}

\item{...}{Not used.}

\item{normalize}{Should statistic be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{sort}{Should results be sorted by the size of the statistic? Default is \code{TRUE}.
Multioutput predictions are sorted by row means.}

\item{top_m}{How many statistics should be shown? By default \code{Inf} (show all).}

\item{eps}{Threshold below which numerator values are set to 0.}
}
\description{
Friedman and Popescu's Pairwise Interaction Strength.
}
\section{Methods (by class)}{
\itemize{
\item \code{H2_pairwise(default)}: Default pairwise interaction strength (not implemented).

\item \code{H2_pairwise(interaction_statistics)}: Pairwise interaction strength from "interaction_statistics".

}}
\references{
Friedman, Jerome H., and Bogdan E. Popescu. "Predictive Learning via Rule Ensembles."
The Annals of Applied Statistics 2, no. 3 (2008): 916-54.
}
