% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2_threeway.R
\name{H2_threeway}
\alias{H2_threeway}
\alias{H2_threeway.default}
\alias{H2_threeway.interact}
\title{Three-way Interaction Strength}
\usage{
H2_threeway(object, ...)

\method{H2_threeway}{default}(object, ...)

\method{H2_threeway}{interact}(
  object,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = 15L,
  eps = 1e-08,
  plot = TRUE,
  fill = "#2b51a1",
  ...
)
}
\arguments{
\item{object}{Object of class "interact".}

\item{...}{Further parameters passed to \code{geom_bar()}.}

\item{normalize}{Should statistic be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{sort}{Should results be sorted by the size of the statistic? Default is \code{TRUE}.
Multioutput predictions are sorted by row means.}

\item{top_m}{How many statistics should be shown? By default \code{15}.
Set to \code{Inf} to show all.}

\item{eps}{Threshold below which numerator values are set to 0.}

\item{plot}{Should results be plotted? Default is \code{TRUE}. Set to \code{FALSE} to get
the results as matrix.}

\item{fill}{Color of bar (only for univariate statistics).}
}
\value{
A "ggplot" object (if \code{plot = TRUE}), or a matrix of statistics
(one row per variable, one column per prediction dimension). If no pairwise
statistics have been calculated, the function returns \code{NULL}.
}
\description{
Friedman and Popescu's statistics of three-way interaction strength extracted from the
result of \code{\link[=interact]{interact()}}, see Details. By default, the results are plotted as a
barplot. Set \code{plot = FALSE} to get a matrix of values instead.
}
\section{Methods (by class)}{
\itemize{
\item \code{H2_threeway(default)}: Default pairwise interaction strength.

\item \code{H2_threeway(interact)}: Pairwise interaction strength from "interact" object.

}}
\examples{
# MODEL 1: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)

# Proportion of joint effect coming from three-way interaction
# (for features with strongest overall interactions)
H2_threeway(inter)
H2_threeway(inter, plot = FALSE)

# Absolute measure as alternative
H2_threeway(inter, normalize = FALSE, squared = FALSE)

# MODEL 2: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
H2_threeway(inter)
}
\references{
Friedman, Jerome H., and Bogdan E. Popescu. \emph{"Predictive Learning via Rule Ensembles."}
The Annals of Applied Statistics 2, no. 3 (2008): 916-54.
}
\seealso{
\code{\link[=interact]{interact()}}, \code{\link[=H2]{H2()}}, \code{\link[=H2_overall]{H2_overall()}}, \code{\link[=H2_pairwise]{H2_pairwise()}}
}
