% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{interact}
\alias{interact}
\alias{interact.default}
\alias{interact.ranger}
\alias{interact.Learner}
\title{Calculate Interaction Statistics}
\usage{
interact(object, ...)

\method{interact}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  pairwise_m = 5L,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  ...
)

\method{interact}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  pairwise_m = 5L,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  ...
)

\method{interact}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  pairwise_m = 5L,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)},
for instance \code{type = "response"} in a \code{\link[=glm]{glm()}} model.}

\item{v}{Vector of feature names.}

\item{X}{A data.frame or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing K >= 1 numeric predictions per row. Its first argument represents the
model \code{object}, its second argument a data structure like \code{X}. Additional arguments
(such as \code{type = "response"} in a GLM) can be passed via \code{...}. The default,
\code{\link[stats:predict]{stats::predict()}}, will work in most cases. Note that column names in a resulting
matrix of predictions will be used as default column names in the results.}

\item{pairwise_m}{Number of features for which pairwise statistics are to be
calculated. The features are selected based on Friedman and Popescu's overall
interaction strength \eqn{H^2_j} (rowwise maximum in the multivariate case).
Set to \code{length(v)} to calculate every pair and to 0 to avoid pairwise calculations.}

\item{n_max}{If \code{X} has more than \code{n_max} rows, a random sample of \code{n_max} rows is
selected from \code{X}. In this case, set a random seed for reproducibility.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{verbose}{Should a progress bar be shown? The default is \code{TRUE}.}
}
\value{
An object of class "interact" containing these elements:
\itemize{
\item \code{f}: Matrix with (centered) predictions \eqn{F}.
\item \code{mean_f2}: (Weighted) column means of \code{f}. Used to normalize most statistics.
\item \code{F_j}: List of matrices, each representing (centered)
partial dependence functions \eqn{F_j}.
\item \code{F_not_j}: List of matrices with (centered) partial dependence
functions \eqn{F_{\setminus j}} of other features.
\item \code{F_jk}: List of matrices, each representing (centered) bivariate
partial dependence functions \eqn{F_{jk}}.
\item \code{w}: Same as input \code{w}.
\item \code{v}: Same as input \code{v}.
\item \code{v_pairwise}: Subset of \code{v} with largest \code{H2_j} used for pairwise calculations.
\item \code{combs}: Named list of variable pairs for which pairwise partial
dependence functions are available.
}
}
\description{
This is the main function of the package. It does the expensive calculations behind
these interaction statistics:
\itemize{
\item Total interaction strength \eqn{H^2}, a statistic measuring the proportion of
prediction variability unexplained by main effects of \code{v}, see \code{\link[=H2]{H2()}} for details.
\item Friedman and Popescu's \eqn{H^2_j} statistic of overall interaction strength per
feature, see \code{\link[=H2_j]{H2_j()}} for details.
\item Friedman and Popescu's \eqn{H^2_{jk}} statistic of pairwise interaction strength,
see \code{\link[=H2_jk]{H2_jk()}} for details.
}

Furthermore, it allows to calculate an experimental partial dependence based
measure of feature importance, \eqn{\textrm{PDI}_j^2}. It equals the proportion of
prediction variability unexplained by other features, see \code{\link[=PDI_j]{PDI_j()}} for details.
(This statistic is not shown by \code{summary()} or \code{plot()}.)

Instead of using \code{summary()}, interaction statistics can also be obtained via the
more flexible functions \code{\link[=H2]{H2()}}, \code{\link[=H2_j]{H2_j()}}, and \code{\link[=H2_jk]{H2_jk()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{interact(default)}: Default interact method.

\item \code{interact(ranger)}: Method for "ranger" models.

\item \code{interact(Learner)}: Method for "mlr3" models.

}}
\examples{
# MODEL 1: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
inter
plot(inter)
summary(inter)
  
# Absolute pairwise interaction strengths
H2_jk(inter, normalize = FALSE, squared = FALSE)

# MODEL 2: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
plot(inter)
summary(inter)

# MODEL 3: Gamma GLM with log link
fit <- glm(Sepal.Length ~ ., data = iris, family = Gamma(link = log))

# No interactions for additive features, at least on link scale
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
summary(inter)

# On original scale, we have interactions everywhere...
inter <- interact(
  fit, v = names(iris[-1]), X = iris, type = "response", verbose = FALSE
)
summary(inter)
}
\references{
Friedman, Jerome H., and Bogdan E. Popescu. \emph{"Predictive Learning via Rule Ensembles."}
The Annals of Applied Statistics 2, no. 3 (2008): 916-54.
}
\seealso{
\code{\link[=H2]{H2()}}, \code{\link[=H2_j]{H2_j()}}, and \code{\link[=H2_jk]{H2_jk()}} for specific statistics calculated from the
resulting object.
}
