% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_profiles.R
\name{pd_profiles}
\alias{pd_profiles}
\alias{pd_profiles.default}
\alias{pd_profiles.ranger}
\alias{pd_profiles.Learner}
\title{Partial Dependence Profiles}
\usage{
pd_profiles(object, ...)

\method{pd_profiles}{default}(
  object,
  v,
  X,
  pred_fun = stats::predict,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  ...
)

\method{pd_profiles}{ranger}(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  ...
)

\method{pd_profiles}{Learner}(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)}, for instance
\code{type = "response"} in a \code{\link[=glm]{glm()}} model.}

\item{v}{Vector of feature names.}

\item{X}{A data.frame or matrix serving as background dataset.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing K >= 1 numeric predictions per row. Its first argument represents the
model \code{object}, its second argument a data structure like \code{X}. Additional arguments
(such as \code{type = "response"} in a GLM) can be passed via \code{...}. The default,
\code{\link[stats:predict]{stats::predict()}}, will work in most cases. Note that column names in a resulting
matrix of predictions will be used as default column names in the results.}

\item{grid}{Named list. Each element specifies the evaluation grid for the
corresponding feature. Missing components are automatically generated via
\code{\link[=univariate_grid]{univariate_grid()}}. If \code{v} has length 1, then \code{grid} can also be a vector.}

\item{grid_size}{Approximate grid size.}

\item{trim}{A non-discrete numeric \code{z} is trimmed at these quantile probabilities
before calculations. Set to \code{c(0, 1)} for no trimming.}

\item{strategy}{How to find evaluation points of non-discrete numeric columns?
Either "quantile" or "uniform" (via \code{\link[=pretty]{pretty()}}), see description of
\code{\link[=univariate_grid]{univariate_grid()}}.}

\item{n_max}{If \code{X} has more than \code{n_max} rows, a random sample of \code{n_max} rows is
selected from \code{X}. In this case, set a random seed for reproducibility.}

\item{w}{Optional vector of case weights for each row of \code{X}.}
}
\value{
A list of PD profiles per variable in \code{v}. Has additional class "pd_profiles".
}
\description{
For a set \code{v} of features, their individual partial dependence is evaluated over
evaluation grids. The grids can be prespecified for any subset of the features.
The function calls the prediction function m times, where m is the number of
features. Multivariate predictions are supported.
}
\section{Methods (by class)}{
\itemize{
\item \code{pd_profiles(default)}: Default method.

\item \code{pd_profiles(ranger)}: Method for "ranger" models.

\item \code{pd_profiles(Learner)}: Method for "mlr3" models.

}}
\examples{
# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ ., data = iris)
pd <- pd_profiles(fit, v = names(iris[-1]), X = iris)
pd$Species
head(pd$Sepal.Width)

pd <- pd_profiles(fit, v = "Petal.Width", X = iris, grid = seq(1, 0, by = -0.5))
pd$Petal.Width 

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
v <- names(iris[3:5])
partial_grid <- list(Petal.Width = seq(0, 1, by = 0.5))
pd <- pd_profiles(fit, v = v, X = iris, grid = partial_grid, verbose = FALSE)
pd$Species
 
# MODEL THREE: Gamma GLM -> pass options to predict() via ...
fit <- glm(
  Sepal.Length ~ . + Petal.Width:Species, 
  data = iris, 
  family = Gamma(link = log)
)
pd_profiles(fit, v = "Species", X = iris, type = "response")$Species
}
\references{
Friedman, Jerome H. \emph{"Greedy Function Approximation: A Gradient Boosting Machine."}
Annals of Statistics 29, no. 5 (2001): 1189-1232.
}
