% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_raw.R
\name{pd_raw}
\alias{pd_raw}
\title{Barebone Partial Dependence (PD) Function}
\usage{
pd_raw(
  object,
  v,
  X,
  grid,
  pred_fun = stats::predict,
  n_max = 1000L,
  w = NULL,
  compress_X = TRUE,
  compress_grid = TRUE,
  check = TRUE,
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{v}{Vector of feature names.}

\item{X}{A data.frame or matrix serving as background dataset.}

\item{grid}{A vector (if \code{length(v) == 1L}), or a matrix/data.frame otherwise.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing K >= 1 numeric predictions per row. Its first argument represents the
model \code{object}, its second argument a data structure like \code{X}. Additional arguments
(such as \code{type = "response"} in a GLM) can be passed via \code{...}. The default,
\code{\link[stats:predict]{stats::predict()}}, will work in most cases. Note that column names in a resulting
matrix of predictions will be used as default column names in the results.}

\item{n_max}{If \code{X} has more than \code{n_max} rows, a random sample of \code{n_max} rows is
selected from \code{X}. In this case, set a random seed for reproducibility.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{compress_X}{If \code{X} has a single non-\code{v} column: should duplicates be removed
and compensated via case weights? Default is \code{TRUE}.}

\item{compress_grid}{Should duplicates in \code{grid} be removed and PDs mapped back to
the original grid index? Default is \code{TRUE}.}

\item{check}{Should input checks be applied? Default is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)}, for instance
\code{type = "response"} in a \code{\link[=glm]{glm()}} model.}
}
\value{
A matrix of partial dependence values (one column per prediction dimension,
one row per grid row, in the same order as \code{grid}).
}
\description{
Workhorse of \code{\link[=pd_profiles]{pd_profiles()}} and \code{\link[=interaction_statistics]{interaction_statistics()}}.
Furthermore, it can be used to calculate multi-dimensional partial dependencies.
In this case, provide your own evaluation grid, for instance via \code{\link[=multivariate_grid]{multivariate_grid()}}.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
pd_raw(fit, v = "Petal.Width", X = iris, pred_fun = predict, grid = 1:2)

# 2D PD
grid <- multivariate_grid(iris[4:5])
grid
values <- pd_raw(fit, X = iris, v = colnames(iris)[4:5], grid = grid)
result <- data.frame(grid, values)
head(result, 2)
}
\references{
Friedman, Jerome H. "Greedy Function Approximation: A Gradient Boosting Machine."
Annals of Statistics 29 (2000): 1189-1232.
}
