% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_raw.R
\name{pd_raw}
\alias{pd_raw}
\title{Barebone Partial Dependence (PD) Function}
\usage{
pd_raw(
  object,
  v,
  X,
  grid,
  pred_fun = stats::predict,
  w = NULL,
  compress_X = TRUE,
  compress_grid = TRUE,
  ...
)
}
\arguments{
\item{object}{Fitted model object.}

\item{v}{Vector of feature names.}

\item{X}{A data.frame or matrix serving as background dataset.}

\item{grid}{Named list. Each element specifies the evaluation grid for the
corresponding feature. Missing components are automatically generated via
\code{\link[=univariate_grid]{univariate_grid()}}. If \code{v} has length 1, then \code{grid} can also be a vector.}

\item{pred_fun}{Prediction function of the form \verb{function(object, X, ...)},
providing K >= 1 numeric predictions per row. Its first argument represents the
model \code{object}, its second argument a data structure like \code{X}. Additional arguments
(such as \code{type = "response"} in a GLM) can be passed via \code{...}. The default,
\code{\link[stats:predict]{stats::predict()}}, will work in most cases. Note that column names in a resulting
matrix of predictions will be used as default column names in the results.}

\item{w}{Optional vector of case weights for each row of \code{X}.}

\item{compress_X}{If \code{X} has a single non-\code{v} column: should duplicates be removed
and compensated via case weights? Default is \code{TRUE}.}

\item{compress_grid}{Should duplicates in \code{grid} be removed and PDs mapped back to
the original grid index? Default is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{pred_fun(object, X, ...)}, for instance
\code{type = "response"} in a \code{\link[=glm]{glm()}} model.}
}
\value{
A matrix of partial dependence values (one column per prediction dimension,
one row per grid row, in the same order as \code{grid}).
}
\description{
Workhorse of the package.
}
\references{
Friedman, Jerome H. \emph{"Greedy Function Approximation: A Gradient Boosting Machine."}
Annals of Statistics 29, no. 5 (2001): 1189-1232.
}
