% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstats.R
\name{plot.hstats}
\alias{plot.hstats}
\title{Plot Method for "hstats" Object}
\usage{
\method{plot}{hstats}(
  x,
  which = 1:2,
  top_m = 15L,
  fill = "#2b51a1",
  facet_scales = "free",
  ncol = 2L,
  rotate_x = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class "hstats".}

\item{which}{Which statistic(s) to be shown? Default is \code{1:2}, i.e., show both
\eqn{H^2_j} (1) and \eqn{H^2_{jk}} (2). To also show three-way interactions,
use \code{1:3}.}

\item{top_m}{Maximum number of rows of results to plot.}

\item{fill}{Color of bars.}

\item{facet_scales}{Value passed to \code{ggplot2::facet_wrap(scales = ...)}.}

\item{ncol}{Passed to \code{ggplot2::facet_wrap()}.}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{...}{Further arguments passed to statistics, e.g., \code{normalize = FALSE}.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for object of class "hstats".
}
\seealso{
See \code{\link[=hstats]{hstats()}} for examples.
}
