% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{plot.hstats_matrix}
\alias{plot.hstats_matrix}
\title{Plots "hstats_matrix" Object}
\usage{
\method{plot}{hstats_matrix}(
  x,
  top_m = 15L,
  multi_output = c("grouped", "facets"),
  fill = "#2b51a1",
  facet_scales = "free",
  ncol = 2L,
  rotate_x = FALSE,
  err_type = c("SE", "SD", "No"),
  ...
)
}
\arguments{
\item{x}{An object of class "hstats_matrix".}

\item{top_m}{How many rows should be plotted? (\code{Inf} for all.)}

\item{multi_output}{How should multi-output models be represented?
Either as "grouped" barplot (the default) or via "facets".}

\item{fill}{Color of bars.}

\item{facet_scales}{Value passed as \code{scales} argument to \verb{[ggplot2::facet_wrap()]}.}

\item{ncol}{Passed to \verb{[ggplot2::facet_wrap()]}.}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{err_type}{The error type to show, by default "SE" (standard errors). Set to
"SD" for standard deviations (SE * sqrt(m_rep)), or "No" for no bars.
Currently, supported only for \code{\link[=perm_importance]{perm_importance()}}.}

\item{...}{Passed to \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}}.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for objects of class "hstats_matrix".
}
