% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{plot.ice}
\alias{plot.ice}
\title{Plots "ice" Object}
\usage{
\method{plot}{ice}(
  x,
  center = FALSE,
  alpha = 0.2,
  color = getOption("hstats.color"),
  facet_scales = "fixed",
  rotate_x = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class "ice".}

\item{center}{Should curves be centered? Default is \code{FALSE}.}

\item{alpha}{Transparency passed to \code{ggplot2::geom_line()}.}

\item{color}{Color of lines and points (in case there is no color/fill aesthetic).
The default equals the global option \code{hstats.color = "#3b528b"}.
To change the global option, use \verb{options(stats.color = new value)}.}

\item{facet_scales}{Value passed as \code{scales} argument to \verb{[ggplot2::facet_wrap()]}.}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{...}{Passed to \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}}.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for objects of class "ice".
}
\seealso{
See \code{\link[=ice]{ice()}} for examples.
}
