% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{plot.interact}
\alias{plot.interact}
\title{Plot Method for "interact" Object}
\usage{
\method{plot}{interact}(x, stat = 1:2, top_m = 15L, fill = "#2b51a1", ...)
}
\arguments{
\item{x}{An object of class "interact".}

\item{stat}{Which statistic(s) to be shown? Default is \code{1:2}, i.e., show both
\eqn{H^2_j} (1) and \eqn{H^2_{jk}} (2).}

\item{top_m}{Maximum number of rows of results to plot.}

\item{fill}{Color of bars (univariate case).}

\item{...}{Further arguments passed to statistics \code{\link[=H2_overall]{H2_overall()}} and \code{\link[=H2_pairwise]{H2_pairwise()}}.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for object of class "interact".
}
\seealso{
See \code{\link[=interact]{interact()}} for examples.
}
