% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dep.R
\name{plot.partial_dep}
\alias{plot.partial_dep}
\title{Plots "partial_dep" Object}
\usage{
\method{plot}{partial_dep}(
  x,
  rotate_x = FALSE,
  color = "#2b51a1",
  facet_scales = "free_y",
  show_points = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class "partial_dep".}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{color}{Color of lines and points (in case there is no color/fill aesthetic).}

\item{facet_scales}{Value passed to \code{facet_wrap(scales = ...)}.}

\item{show_points}{Logical flag indicating whether to show points (default) or not.}

\item{...}{Arguments passed to geometries.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for objects of class "partial_dep". Can do (grouped) line plots or
heatmaps.
}
\seealso{
See \code{\link[=partial_dep]{partial_dep()}} for examples.
}
