% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dep.R
\name{plot.partial_dep}
\alias{plot.partial_dep}
\title{Plots "partial_dep" Object}
\usage{
\method{plot}{partial_dep}(
  x,
  color = getOption("hstats.color"),
  facet_scales = "free_y",
  rotate_x = FALSE,
  show_points = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class "partial_dep".}

\item{color}{Color of lines and points (in case there is no color/fill aesthetic).
The default equals the global option \code{hstats.color = "#3b528b"}.
To change the global option, use \verb{options(stats.color = new value)}.}

\item{facet_scales}{Value passed as \code{scales} argument to \verb{[ggplot2::facet_wrap()]}.}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{show_points}{Logical flag indicating whether to show points (default) or not.}

\item{...}{Arguments passed to geometries.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for objects of class "partial_dep". Can do (grouped) line plots or
heatmaps.
}
\seealso{
See \code{\link[=partial_dep]{partial_dep()}} for examples.
}
