% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{plot.pd}
\alias{plot.pd}
\title{PDP}
\usage{
\method{plot}{pd}(x, rotate_x = FALSE, color = "#2b51a1", facet_scales = "free_y", ...)
}
\arguments{
\item{x}{An object of class "pd".}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{color}{Color of lines and points.}

\item{facet_scales}{Value passed to \code{facet_wrap(scales = ...)}, only relevant
for multivariate output.}

\item{...}{Further arguments passed to geometries.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for object of class "pd". It supports multivariate predictions and a
BY variable, but only univariable \code{v}.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  fit <- lm(Sepal.Length ~ . + Species * Petal.Length, data = iris)
  plot(partial_dep(fit, v = "Species", X = iris))

  # Multi-response linear regression
  fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
  
  # BY = iris$Species also works. This allows to use transformed features.
  pd <- partial_dep(fit, v = "Petal.Width", X = iris, BY = "Species", grid_size = 13)
  plot(pd)
}

}
\seealso{
\code{\link[=partial_dep]{partial_dep()}}
}
