% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_importance.R
\name{plot.perm_importance}
\alias{plot.perm_importance}
\title{Plots "perm_importance" Object}
\usage{
\method{plot}{perm_importance}(
  x,
  normalize = FALSE,
  sort = TRUE,
  top_m = 15L,
  agg_cols = c("sum", "mean", "max", "no"),
  err_type = c("se", "sd", "no"),
  facet_scales = "fixed",
  ncol = 2L,
  rotate_x = FALSE,
  fill = "#2b51a1",
  ...
)
}
\arguments{
\item{x}{An object of class "perm_importance".}

\item{normalize}{Should importance statistics be divided by performance?
Default is \code{FALSE}. Normalization is applied after collapsing multivariate losses.}

\item{sort}{Should results be sorted by importance? Default is \code{TRUE}.
The multivariate case is sorted by row sums.}

\item{top_m}{Of how many features should importance values be shown?
Default is \code{Inf}, i.e., show all.}

\item{agg_cols}{Name of function used to collapse multivariate losses.
Default is \code{"sum"}. Set to \code{"no"} for no aggregation. Other options are "mean" and
"max". This step is applied before normalization.}

\item{err_type}{Should standard errors ("se", default) or standard deviations ("sd")
be calculated? Only relevant if \code{perms > 1}. Set to "no" to hide errors.}

\item{facet_scales}{Value passed to \code{ggplot2::facet_wrap(scales = ...)}.}

\item{ncol}{Passed to \code{ggplot2::facet_wrap()}.}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{fill}{Color of bars.}

\item{...}{Arguments passed to \link[ggplot2:geom_bar]{ggplot2::geom_bar}.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for objects of class "perm_importance".
}
\seealso{
See \code{\link[=perm_importance]{perm_importance()}} for examples.
}
