% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_importance.R
\name{plot.perm_importance}
\alias{plot.perm_importance}
\title{Plots "perm_importance" Object}
\usage{
\method{plot}{perm_importance}(
  x,
  top_m = 15L,
  err_type = c("se", "sd", "no"),
  multi_output = c("dodge", "facets"),
  facet_scales = "fixed",
  ncol = 2L,
  rotate_x = FALSE,
  fill = "#2b51a1",
  ...
)
}
\arguments{
\item{x}{An object of class "perm_importance".}

\item{top_m}{How many rows should be shown? (\code{Inf} to show all.)}

\item{err_type}{The error type to show, by default "se" (standard errors). Set to
"sd" for standard deviations (se * sqrt(perms)), or "no" for no bars.}

\item{multi_output}{How should multi-output models be represented?
Either "dodge" (dodged bars, the default) or via "facets".}

\item{facet_scales}{Value passed to \code{ggplot2::facet_wrap(scales = ...)}.}

\item{ncol}{Passed to \code{ggplot2::facet_wrap()}.}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{fill}{Color of bar (only for univariate statistics).}

\item{...}{Arguments passed to \link[ggplot2:geom_bar]{ggplot2::geom_bar}.}
}
\value{
An object of class "ggplot".
}
\description{
Plot method for objects of class "perm_importance".
}
\seealso{
See \code{\link[=perm_importance]{perm_importance()}} for examples.
}
