% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_core.R
\name{print.pd}
\alias{print.pd}
\title{PD Print}
\usage{
\method{print}{pd}(x, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Currently unused.}
}
\value{
Invisibly, \code{x} is returned.
}
\description{
Print function for result of \code{\link[=pd]{pd()}}.
}
\seealso{
\code{\link[=pd]{pd()}}
}
