% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstats.R
\name{summary.hstats}
\alias{summary.hstats}
\title{Summary Method}
\usage{
\method{summary}{hstats}(
  object,
  normalize = TRUE,
  squared = TRUE,
  sort = TRUE,
  top_m = 6L,
  eps = 1e-08,
  ...
)
}
\arguments{
\item{object}{Object of class "hstats".}

\item{normalize}{Should statistics be normalized? Default is \code{TRUE}.}

\item{squared}{Should \emph{squared} statistics be returned? Default is \code{TRUE}.}

\item{sort}{Should results be sorted? Default is \code{TRUE}.
(Multioutput is sorted by row means.)}

\item{top_m}{How many rows should be shown? (\code{Inf} to show all.)}

\item{eps}{Threshold below which numerator values are set to 0.}

\item{...}{Currently not used.}
}
\value{
A named list of statistics.
}
\description{
Summary method for "hstats" object.
}
\seealso{
See \code{\link[=hstats]{hstats()}} for examples.
}
