% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstats.R
\name{summary.hstats}
\alias{summary.hstats}
\title{Summary Method}
\usage{
\method{summary}{hstats}(object, top_m = 6L, ...)
}
\arguments{
\item{object}{An object of class "hstats".}

\item{top_m}{Maximum number of rows of results to print.}

\item{...}{Further arguments passed to statistics, e.g., \code{normalize = FALSE}.}
}
\value{
A named list of statistics.
}
\description{
Summary method for "hstats" object.
}
\seealso{
See \code{\link[=hstats]{hstats()}} for examples.
}
