% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{summary.interact}
\alias{summary.interact}
\title{Summary Method}
\usage{
\method{summary}{interact}(object, top_m = 6L, ...)
}
\arguments{
\item{object}{An object of class "interact".}

\item{top_m}{Maximum number of rows of results to print.}

\item{...}{Further arguments passed to statistics, e.g., \code{normalize = FALSE}.}
}
\value{
A named list of statistics.
}
\description{
Summary method for "interact" object.
}
\seealso{
See \code{\link[=interact]{interact()}} for examples.
}
